+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod backend;
 pub mod check_hash;
 pub mod constraints;
 pub mod miller_rabin;
